# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ModelInfo {
  id: string;
  name: string;
  group: string;
}

export interface ModelListResult {
  models: ModelInfo[];
  default: string;
}

export interface VoiceProfileResult {
  avgClarity: number;
  avgConfidence: number;
  avgConciseness: number;
  avgConnection: number;
  totalSessions: number;
}

export interface TranscribeRequest {
  /** Base64-encoded audio data */
  audioBase64: string;
  /** MIME type of the audio */
  mimeType?: string;
  /** Actual recording duration in seconds (measured client-side) */
  durationSeconds?: number;
  /** Base64-encoded JPEG video frames for body language analysis */
  videoFrames?: string[];
}

export interface TranscriptionResult {
  transcript: string;
  /** HTML transcript with filler/hedge words highlighted */
  highlightedTranscript?: string;
  durationSeconds: number;
  wordCount: number;
  wpm: number;
  /** Number of filler/hedge words detected */
  fillerCount?: number;
  /** AI analysis of body language from video frames */
  bodyLanguageAnalysis?: string;
}

export interface EvaluateRequest {
  transcript: string;
  promptLabel?: string;
  promptText?: string;
  /** Body language analysis to incorporate into feedback */
  bodyLanguageAnalysis?: string;
  /** Model ID to use for evaluation (defaults to Claude) */
  modelId?: string;
}

export interface EvaluationResult {
  feedback: string;
  modelUsed?: string;
}

export interface SpeakRequest {
  /** Text to convert to speech */
  text: string;
}

export interface SpeakResult {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
}

export interface SaveSessionRequest {
  promptLabel?: string;
  promptText?: string;
  transcript: string;
  highlightedTranscript?: string;
  wordCount: number;
  wpm: number;
  durationSeconds: number;
  feedback?: string;
  bodyLanguageAnalysis?: string;
  modelUsed?: string;
  clarityScore?: number;
  confidenceScore?: number;
  concisenessScore?: number;
  connectionScore?: number;
  fillerCount?: number;
}

export interface SaveSessionResult {
  id: string;
  createdAt: string;
}

export interface SessionSummary {
  id: string;
  promptLabel?: string;
  wpm: number;
  wordCount: number;
  durationSeconds: number;
  createdAt: string;
  transcript: string;
  highlightedTranscript?: string;
  feedback?: string;
  bodyLanguageAnalysis?: string;
  modelUsed?: string;
  clarityScore?: number;
  confidenceScore?: number;
  concisenessScore?: number;
  connectionScore?: number;
  fillerCount?: number;
}

export interface SessionListResult {
  sessions: SessionSummary[];
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};