# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transcribe:
    post:
      operationId: transcribeAudio
      tags: [coach]
      summary: Transcribe audio and optionally analyze video frames
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeRequest"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /evaluate:
    post:
      operationId: evaluateTranscript
      tags: [coach]
      summary: Evaluate transcript with AI coaching
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Evaluation feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /speak:
    post:
      operationId: speakFeedback
      tags: [coach]
      summary: Convert text feedback to spoken audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakRequest"
      responses:
        "200":
          description: Audio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions:
    get:
      operationId: getSessions
      tags: [coach]
      summary: Get practice session history for the current user
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveSession
      tags: [coach]
      summary: Save a practice session
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSessionRequest"
      responses:
        "200":
          description: Saved session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveSessionResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /voice-profile:
    get:
      operationId: getVoiceProfile
      tags: [coach]
      summary: Get the user's rolling Voice Fingerprint averages
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Voice profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceProfileResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    VoiceProfileResult:
      type: object
      properties:
        avgClarity:
          type: number
        avgConfidence:
          type: number
        avgConciseness:
          type: number
        avgConnection:
          type: number
        totalSessions:
          type: integer
      required:
        - avgClarity
        - avgConfidence
        - avgConciseness
        - avgConnection
        - totalSessions
    TranscribeRequest:
      type: object
      properties:
        audioBase64:
          type: string
          description: Base64-encoded audio data
        mimeType:
          type: string
          description: MIME type of the audio
        durationSeconds:
          type: number
          description: Actual recording duration in seconds (measured client-side)
        videoFrames:
          type: array
          items:
            type: string
          description: Base64-encoded JPEG video frames for body language analysis
      required:
        - audioBase64
    TranscriptionResult:
      type: object
      properties:
        transcript:
          type: string
        highlightedTranscript:
          type: string
          description: HTML transcript with filler/hedge words highlighted
        durationSeconds:
          type: number
        wordCount:
          type: integer
        wpm:
          type: integer
        fillerCount:
          type: integer
          description: Number of filler/hedge words detected
        bodyLanguageAnalysis:
          type: string
          description: AI analysis of body language from video frames
      required:
        - transcript
        - durationSeconds
        - wordCount
        - wpm
    EvaluateRequest:
      type: object
      properties:
        transcript:
          type: string
        promptLabel:
          type: string
        promptText:
          type: string
        bodyLanguageAnalysis:
          type: string
          description: Body language analysis to incorporate into feedback
        modelId:
          type: string
          description: Model ID to use for evaluation (defaults to Claude)
      required:
        - transcript
    EvaluationResult:
      type: object
      properties:
        feedback:
          type: string
        modelUsed:
          type: string
      required:
        - feedback
    SpeakRequest:
      type: object
      properties:
        text:
          type: string
          description: Text to convert to speech
      required:
        - text
    SpeakResult:
      type: object
      properties:
        audioBase64:
          type: string
          description: Base64-encoded MP3 audio
      required:
        - audioBase64
    SaveSessionRequest:
      type: object
      properties:
        promptLabel:
          type: string
        promptText:
          type: string
        transcript:
          type: string
        highlightedTranscript:
          type: string
        wordCount:
          type: integer
        wpm:
          type: integer
        durationSeconds:
          type: integer
        feedback:
          type: string
        bodyLanguageAnalysis:
          type: string
        modelUsed:
          type: string
        clarityScore:
          type: number
        confidenceScore:
          type: number
        concisenessScore:
          type: number
        connectionScore:
          type: number
        fillerCount:
          type: integer
      required:
        - transcript
        - wordCount
        - wpm
        - durationSeconds
    SaveSessionResult:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
      required:
        - id
        - createdAt
    SessionListResult:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
      required:
        - sessions
    SessionSummary:
      type: object
      properties:
        id:
          type: string
        promptLabel:
          type: string
        wpm:
          type: integer
        wordCount:
          type: integer
        durationSeconds:
          type: integer
        createdAt:
          type: string
        transcript:
          type: string
        highlightedTranscript:
          type: string
        feedback:
          type: string
        bodyLanguageAnalysis:
          type: string
        modelUsed:
          type: string
        clarityScore:
          type: number
        confidenceScore:
          type: number
        concisenessScore:
          type: number
        connectionScore:
          type: number
        fillerCount:
          type: integer
      required:
        - id
        - wpm
        - wordCount
        - durationSeconds
        - createdAt
        - transcript
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...