# lib/api-zod/src/generated/types/evaluateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluateRequest {
  transcript: string;
  promptLabel?: string;
  promptText?: string;
  /** Body language analysis to incorporate into feedback */
  bodyLanguageAnalysis?: string;
  /** Model ID to use for evaluation (defaults to Claude) */
  modelId?: string;
}